+++ examplepatches/hackingguy/morphe-patches/patches/src/main/kotlin/app/morphe/patches/mygate/premium/Fingerprints.kt
@@ -64,7 +64,14 @@ internal object ShowUpgradeDialogFingerprint : Fingerprint(
     definingClass = "Lcom/mygate/user/utilities/KotlinUtils\$Companion;",
     returnType = "V",
     parameters = listOf("Landroidx/fragment/app/FragmentActivity;"),
-    strings = listOf("FeatureNavigation")
+    filters = listOf(
+        app.morphe.patcher.methodCall(
+            definingClass = "Landroid/app/Dialog;",
+            name = "show",
+            returnType = "V",
+            parameters = emptyList()
+        )
+    )
 )
 
 /**